refactor: follow project module refactor
