+++ src/follow-project/follow-project.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FollowProjectService } from './service/follow-project/follow-project.service';
-import { FollowProjectController } from './controller/follow-project/follow-project.controller';
+import { FollowProjectService } from './service/follow-project.service';
+import { FollowProjectController } from './controller/follow-project.controller';
 import { SharedServicesModule } from 'src/shared/shared-services.module';
 
 @Module({
